Refactor-6 Move struct definations to definitions
